Fix bugs writing out variables
There were two problems with the code:

 - pt and eta were defined in the list of output
   variables. Unfortunately these (like all 4 momentum variables)
   can't be accessed via the normal string accessors and are hard
   coded instead in the writers. Defining them in both places throws
   an exception when HDF5 tries to add the same variable twice.

 - We were trying to read truth labels out of the b-tagging container
   in the jet. These are actually jet attributes, so accessing them is
   slightly different.

I also split the config structures into one for tracks and another one
for jets. Given that some of the fields only make sense in one case
(i.e. truth labels in the jet case, length in the track case) using
the same struct for both was a hack anyway.